djangogirls/requirements.txt
django
djangorestframework
coreapi
Markdown
django-filter
django-crispy-forms
django-guardian